requirements.txt
streamlit>=1.30.0
streamlit-webrtc>=0.47.0
audio-recorder-streamlit>=0.0.8
groq>=0.4.0
openai-whisper>=20231117
gTTS>=2.4.0
bcrypt>=4.1.2
pydub>=0.25.1
python-dotenv>=1.0.0
